Introduced a dedicated hero image utility that uses import.meta.glob to map hero image filenames to URLs for reuse across the app
