+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export * from './env.js';
+export * from './heroImages.js';
 export * from './langUtils.js';
 export * from './storage.js';
 export { default as storage } from './storage.js';
